Change atoi() to atol()

+++ Assignment 1/Part1/double.c
@@ -4,7 +4,7 @@
 #include <unistd.h>
 
 int main(int argc, char *argv[]) {
-	unsigned long inputNum = atoi(argv[argc - 1]);
+	unsigned long inputNum = atol(argv[argc - 1]);
 	unsigned long result = 2 * inputNum;
 
 	if (argc == 2) {
